Refactor ExposuresViewWrapper
- Split the existing hybrid view into two separate views
- Added new LocalizedView protocol for easier access to localizations
- Extracted the related components to separate files
