+++ Koronavilkku/Koronavilkku/Localization/en.lproj/Localizable.strings
@@ -50,29 +50,28 @@
 "SymptomsElement.Text.Title" = "Do you think you may have COVID-19?";
 "SymptomsElement.Text.Body" = "Take an online medical COVID-19 symptom check-up";
 
-// MARK: ExposureViewController
+// MARK: ExposuresViewController
 "ExposuresViewController.Text.TitleHasExposures" = "Potential exposure";
 
-// MARK: ExposuresViewWrapper
-"ExposuresViewWrapper.HasExposureText.Heading" = "In the past 14 days, you have been close to a person who has reported a COVID-19 infection.";
-"ExposuresViewWrapper.HasExposureText.ContactCardTitle" = "Contact healthcare services";
-"ExposuresViewWrapper.HasExposureText.ContactCardText" = "Information about your potential exposure will only be communicated to healthcare services if you contact them yourself.";
-"ExposuresViewWrapper.HasExposureText.ContactButtonTitle" = "Contact";
-"ExposuresViewWrapper.HasExposureText.ExposuresButtonTitle" = "Show exposure notifications";
-"ExposuresViewWrapper.HasExposureText.InstructionsTitle" = "What should I do?";
-"ExposuresViewWrapper.HasExposureText.InstructionsLede" = "Follow these instructions for 14 days:";
-"ExposuresViewWrapper.HasExposureText.InstructionsSymptoms" = "Monitor your health. If you develop any symptom consistent with COVID-19, contact your municipal healthcare service and get tested as soon as possible. Symptoms consistent with COVID-19 are change in sense of smell or taste, shortness of breath, cough, sore throat, runny nose, fever, headaches, muscle pain, unexplained diarrhoea or abdominal pain.";
-"ExposuresViewWrapper.HasExposureText.InstructionsDistancing" = "Avoid meeting other people.";
-"ExposuresViewWrapper.HasExposureText.InstructionsHygiene" = "Maintain good hand hygiene.";
-"ExposuresViewWrapper.HasExposureText.InstructionsAvoidTravel" = "Avoid public places and public transport. If you have to visit public places, keep at least a two-meter distance to other people. If you cannot keep a sufficient distance, use a face mask.";
-"ExposuresViewWrapper.HasExposureText.InstructionsShopping" = "If possible, let someone else handle food shopping and other errands outside the home.";
-"ExposuresViewWrapper.HasExposureText.InstructionsCoughing" = "If you cough or sneeze, use a disposable tissue to cover your mouth and nose. Put the used tissue in a waste bin and wash your hands with water and soap or use hand sanitiser. If you do not have a tissue, cough or sneeze into your sleeve.";
-"ExposuresViewWrapper.HasExposureText.InstructionsRemoteWork" = "Work remotely if possible.";
-"ExposuresViewWrapper.HasExposureText.InstructionsWhenAbroad" = "If you are not in Finland, follow any instructions given by local authorities.";
+"HasExposuresView.Text.Heading" = "In the past 14 days, you have been close to a person who has reported a COVID-19 infection.";
+"HasExposuresView.Text.ContactCardTitle" = "Contact healthcare services";
+"HasExposuresView.Text.ContactCardText" = "Information about your potential exposure will only be communicated to healthcare services if you contact them yourself.";
+"HasExposuresView.Text.ContactButtonTitle" = "Contact";
+"HasExposuresView.Text.ExposuresButtonTitle" = "Show exposure notifications";
+"HasExposuresView.Text.InstructionsTitle" = "What should I do?";
+"HasExposuresView.Text.InstructionsLede" = "Follow these instructions for 14 days:";
+"HasExposuresView.Text.InstructionsSymptoms" = "Monitor your health. If you develop any symptom consistent with COVID-19, contact your municipal healthcare service and get tested as soon as possible. Symptoms consistent with COVID-19 are change in sense of smell or taste, shortness of breath, cough, sore throat, runny nose, fever, headaches, muscle pain, unexplained diarrhoea or abdominal pain.";
+"HasExposuresView.Text.InstructionsDistancing" = "Avoid meeting other people.";
+"HasExposuresView.Text.InstructionsHygiene" = "Maintain good hand hygiene.";
+"HasExposuresView.Text.InstructionsAvoidTravel" = "Avoid public places and public transport. If you have to visit public places, keep at least a two-meter distance to other people. If you cannot keep a sufficient distance, use a face mask.";
+"HasExposuresView.Text.InstructionsShopping" = "If possible, let someone else handle food shopping and other errands outside the home.";
+"HasExposuresView.Text.InstructionsCoughing" = "If you cough or sneeze, use a disposable tissue to cover your mouth and nose. Put the used tissue in a waste bin and wash your hands with water and soap or use hand sanitiser. If you do not have a tissue, cough or sneeze into your sleeve.";
+"HasExposuresView.Text.InstructionsRemoteWork" = "Work remotely if possible.";
+"HasExposuresView.Text.InstructionsWhenAbroad" = "If you are not in Finland, follow any instructions given by local authorities.";
 
-"ExposuresViewWrapper.NoExposuresText.Heading" = "No observed exposures";
-"ExposuresViewWrapper.NoExposuresText.Subtitle" = "Koronavilkku has not detected exposures to the coronavirus in the last 14 days.";
-"ExposuresViewWrapper.NoExposuresText.DisclaimerText" = "However, the application is not able to identify every situation where exposure may occur.";
+"NoExposuresView.Text.Heading" = "No observed exposures";
+"NoExposuresView.Text.Subtitle" = "Koronavilkku has not detected exposures to the coronavirus in the last 14 days.";
+"NoExposuresView.Text.DisclaimerText" = "However, the application is not able to identify every situation where exposure may occur.";
 
 "ExposuresLastCheckedView.Text.LastCheckedAt" = "Last checked %@";
 "ExposuresLastCheckedView.Text.MomentAgo" = "a moment ago";
