+++ Koronavilkku/Koronavilkku/Localization/sv.lproj/Localizable.strings
@@ -50,29 +50,28 @@
 "SymptomsElement.Text.Title" = "Misstänker du smitta?";
 "SymptomsElement.Text.Body" = "Gör en symtombedömning";
 
-// MARK: ExposureViewController
+// MARK: ExposuresViewController
 "ExposuresViewController.Text.TitleHasExposures" = "Eventuell exponering";
 
-// MARK: ExposuresViewWrapper
-"ExposuresViewWrapper.HasExposureText.Heading" = "Du har under de senaste 14 dygnen varit i närheten av en person som meddelat om sin coronasmitta via applikationen.";
-"ExposuresViewWrapper.HasExposureText.ContactCardTitle" = "Kontakta hälsovården";
-"ExposuresViewWrapper.HasExposureText.ContactCardText" = "Information om eventuell exponering förmedlas till hälsovården endast om du själv tar kontakt.";
-"ExposuresViewWrapper.HasExposureText.ContactButtonTitle" = "Ta kontakt";
-"ExposuresViewWrapper.HasExposureText.ExposuresButtonTitle" = "Visa exponerings\U00ADmeddelandena";
-"ExposuresViewWrapper.HasExposureText.InstructionsTitle" = "Gör så här";
-"ExposuresViewWrapper.HasExposureText.InstructionsLede" = "Följ dessa anvisningar i 14 dagar:";
-"ExposuresViewWrapper.HasExposureText.InstructionsSymptoms" = "Var uppmärksam på hur du mår. Om du har ens ett av symtomen som passar in på coronavirus\U00ADinfektion, ta så snabbt som möjligt kontakt med hälsovården i din kommun för att bli testad. Symtom som kan tyda på coronavirus\U00ADinfektion är förändringar i lukt- eller smaksinnet, andningssvårigheter, hosta, halsont, snuva, feber, huvudvärk, muskelvärk samt diarré eller magont utan annan uppenbar orsak.";
-"ExposuresViewWrapper.HasExposureText.InstructionsDistancing" = "Undvik att träffa andra människor.";
-"ExposuresViewWrapper.HasExposureText.InstructionsHygiene" = "Var noggrann med din handhygien.";
-"ExposuresViewWrapper.HasExposureText.InstructionsAvoidTravel" = "Undvik offentliga platser och kollektivtrafik. Om du är tvungen att vistas på offentliga platser ska du hålla ett avstånd på minst två meter till andra människor. Om du inte kan hålla ett tillräckligt långt avstånd, använd ansiktsmask.";
-"ExposuresViewWrapper.HasExposureText.InstructionsShopping" = "Låt en annan person handla mat och sköta andra ärenden utanför hemmet, om möjligt.";
-"ExposuresViewWrapper.HasExposureText.InstructionsCoughing" = "Om du hostar eller nyser ska du skydda din mun och näsa med en engångsnäsduk. Kasta den använda näsduken i soporna och tvätta dina händer med tvål och vatten eller använd hand\U00ADdesinfektions\U00ADmedel. Om du inte har en näsduk till hands, hosta eller nys i armvecket.";
-"ExposuresViewWrapper.HasExposureText.InstructionsRemoteWork" = "Arbeta på distans, om möjligt.";
-"ExposuresViewWrapper.HasExposureText.InstructionsWhenAbroad" = "Om du inte befinner dig i Finland ska du följa de lokala myndigheternas anvisningar.";
+"HasExposuresView.Text.Heading" = "Du har under de senaste 14 dygnen varit i närheten av en person som meddelat om sin coronasmitta via applikationen.";
+"HasExposuresView.Text.ContactCardTitle" = "Kontakta hälsovården";
+"HasExposuresView.Text.ContactCardText" = "Information om eventuell exponering förmedlas till hälsovården endast om du själv tar kontakt.";
+"HasExposuresView.Text.ContactButtonTitle" = "Ta kontakt";
+"HasExposuresView.Text.ExposuresButtonTitle" = "Visa exponerings\U00ADmeddelandena";
+"HasExposuresView.Text.InstructionsTitle" = "Gör så här";
+"HasExposuresView.Text.InstructionsLede" = "Följ dessa anvisningar i 14 dagar:";
+"HasExposuresView.Text.InstructionsSymptoms" = "Var uppmärksam på hur du mår. Om du har ens ett av symtomen som passar in på coronavirus\U00ADinfektion, ta så snabbt som möjligt kontakt med hälsovården i din kommun för att bli testad. Symtom som kan tyda på coronavirus\U00ADinfektion är förändringar i lukt- eller smaksinnet, andningssvårigheter, hosta, halsont, snuva, feber, huvudvärk, muskelvärk samt diarré eller magont utan annan uppenbar orsak.";
+"HasExposuresView.Text.InstructionsDistancing" = "Undvik att träffa andra människor.";
+"HasExposuresView.Text.InstructionsHygiene" = "Var noggrann med din handhygien.";
+"HasExposuresView.Text.InstructionsAvoidTravel" = "Undvik offentliga platser och kollektivtrafik. Om du är tvungen att vistas på offentliga platser ska du hålla ett avstånd på minst två meter till andra människor. Om du inte kan hålla ett tillräckligt långt avstånd, använd ansiktsmask.";
+"HasExposuresView.Text.InstructionsShopping" = "Låt en annan person handla mat och sköta andra ärenden utanför hemmet, om möjligt.";
+"HasExposuresView.Text.InstructionsCoughing" = "Om du hostar eller nyser ska du skydda din mun och näsa med en engångsnäsduk. Kasta den använda näsduken i soporna och tvätta dina händer med tvål och vatten eller använd hand\U00ADdesinfektions\U00ADmedel. Om du inte har en näsduk till hands, hosta eller nys i armvecket.";
+"HasExposuresView.Text.InstructionsRemoteWork" = "Arbeta på distans, om möjligt.";
+"HasExposuresView.Text.InstructionsWhenAbroad" = "Om du inte befinner dig i Finland ska du följa de lokala myndigheternas anvisningar.";
 
-"ExposuresViewWrapper.NoExposuresText.Heading" = "Ingen exponering har upptäckts";
-"ExposuresViewWrapper.NoExposuresText.Subtitle" = "Coronablinkern har inte observerat någon exponering för coronaviruset under de senaste 14 dagarna.";
-"ExposuresViewWrapper.NoExposuresText.DisclaimerText" = "Appen kan emellertid inte identifiera alla eventuella exponeringar.";
+"NoExposuresView.Text.Heading" = "Ingen exponering har upptäckts";
+"NoExposuresView.Text.Subtitle" = "Coronablinkern har inte observerat någon exponering för coronaviruset under de senaste 14 dagarna.";
+"NoExposuresView.Text.DisclaimerText" = "Appen kan emellertid inte identifiera alla eventuella exponeringar.";
 
 "ExposuresLastCheckedView.Text.LastCheckedAt" = "Senast granskad %@";
 "ExposuresLastCheckedView.Text.MomentAgo" = "för ett ögonblick sedan";
