+++ Koronavilkku/Koronavilkku/Localization/sv.lproj/Localizable.stringsdict
@@ -2,7 +2,7 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>ExposuresViewWrapper.HasExposureText.ExposuresButtonNotificationCount</key>
+	<key>HasExposuresView.Text.ExposuresButtonNotificationCount</key>
 	<dict>
 		<key>NSStringLocalizedFormatKey</key>
 		<string>%#@notificationCount@</string>
