+++ Koronavilkku/Koronavilkku/ViewController/ExposuresViewController.swift
@@ -3,6 +3,17 @@ import UIKit
 import SnapKit
 import Combine
 
+protocol ExposuresViewDelegate: AnyObject {
+    func showHowItWorks()
+    func makeContact()
+    func startManualCheck()
+    func showNotificationList()
+}
+
+class ExposuresView : UIView {
+    weak var delegate: ExposuresViewDelegate?
+}
+
 class ExposuresViewController: UIViewController {
     
     enum Text : String, Localizable {
@@ -10,14 +21,17 @@ class ExposuresViewController: UIViewController {
     }
 
     private let exposureRepository: ExposureRepository
-    private let exposuresViewWrapper = ExposuresViewWrapper()
     private var updateTasks = Set<AnyCancellable>()
-    
+    private var exposuresView: ExposuresView!
+    private var containerView: UIView!
+        
     private var exposureStatus: ExposureStatus? {
         didSet {
-            guard let exposureStatus = exposureStatus, exposureStatus != oldValue else { return }
+            guard let exposureStatus = exposureStatus,
+                  exposureStatus != oldValue else { return }
+            
             self.updateNavigationBar()
-            self.exposuresViewWrapper.exposureStatus = exposureStatus
+            self.render()
         }
     }
     
@@ -44,7 +58,6 @@ class ExposuresViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        exposuresViewWrapper.delegate = self
         initUI()
         bindViewModel()
     }
@@ -73,13 +86,17 @@ class ExposuresViewController: UIViewController {
         exposureRepository.detectionStatus()
             .receive(on: RunLoop.main)
             .sink { [weak self] detectionStatus in
-                self?.exposuresViewWrapper.detectionStatus = detectionStatus
+                if let noExposuresView = self?.exposuresView as? NoExposuresView {
+                    noExposuresView.detectionStatus = detectionStatus
+                }
             }
             .store(in: &updateTasks)
         
         exposureRepository.timeFromLastCheck()
             .sink { [weak self] time in
-                self?.exposuresViewWrapper.timeFromLastCheck = time
+                if let noExposuresView = self?.exposuresView as? NoExposuresView {
+                    noExposuresView.timeFromLastCheck = time
+                }
             }
             .store(in: &updateTasks)
     }
@@ -111,12 +128,25 @@ class ExposuresViewController: UIViewController {
     }
 
     private func initUI() {
-        view.removeAllSubviews()
-        
         let margins = UIEdgeInsets(top: 30, left: 20, bottom: 30, right: 20)
-        let contentView = view.addScrollableContentView(backgroundColor: UIColor.Secondary.blueBackdrop, margins: margins)
-        contentView.addSubview(exposuresViewWrapper)
-        exposuresViewWrapper.snp.makeConstraints { make in
+        
+        containerView = view.addScrollableContentView(backgroundColor: UIColor.Secondary.blueBackdrop, margins: margins)
+        
+        render()
+    }
+    
+    private func render() {
+        containerView.removeAllSubviews()
+        
+        if case .exposed(let notificationCount) = exposureStatus {
+            exposuresView = HasExposuresView(notificationCount: notificationCount)
+        } else {
+            exposuresView = NoExposuresView()
+        }
+        
+        exposuresView.delegate = self
+        containerView.addSubview(exposuresView)
+        exposuresView.snp.makeConstraints { make in
             make.edges.equalToSuperview()
         }
     }
