+++ Koronavilkku/KoronavilkkuTests/Localization/LocalizationTests.swift
@@ -10,8 +10,8 @@ class LocalizationTests: XCTestCase {
         verifyTranslations(from: StatusHeaderView.Text.self)
         verifyTranslations(from: ExposuresElement.Text.self)
         verifyTranslations(from: ExposuresViewController.Text.self)
-        verifyTranslations(from: ExposuresViewWrapper.HasExposureText.self)
-        verifyTranslations(from: ExposuresViewWrapper.NoExposuresText.self)
+        verifyTranslations(from: HasExposuresView.Text.self)
+        verifyTranslations(from: NoExposuresView.Text.self)
         verifyTranslations(from: ReportInfectionViewController.Text.self)
         verifyTranslations(from: ChooseCountriesViewController.Text.self)
         verifyTranslations(from: ChooseDestinationViewController.Text.self)
